feat: switch everything to tabs
